+++ src/SitemapBundle.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Ecourty\SitemapBundle;
 
-use Ecourty\SitemapBundle\DependencyInjection\EcourtySitemapExtension;
+use Ecourty\SitemapBundle\DependencyInjection\SitemapExtension;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -18,7 +18,7 @@ class SitemapBundle extends Bundle
     public function getContainerExtension(): ?ExtensionInterface
     {
         if (null === $this->extension) {
-            $this->extension = new EcourtySitemapExtension();
+            $this->extension = new SitemapExtension();
         }
 
         return $this->extension instanceof ExtensionInterface ? $this->extension : null;